Sync20191008

+++ src/serializer/buf_ptr.cc
@@ -20,9 +20,9 @@ buf_ptr_t buf_ptr_t::alloc_zeroed(block_size_t size) {
     return ret;
 }
 
-scoped_device_block_aligned_ptr_t<ser_buffer_t> help_allocate_copy(const ser_buffer_t *copyee,
-                                                                   size_t amount_to_copy,
-                                                                   size_t reserved_size) {
+scoped_device_block_aligned_ptr_t<ser_buffer_t>
+help_allocate_copy(const ser_buffer_t *copyee, size_t amount_to_copy,
+                   size_t reserved_size) {
     rassert(amount_to_copy <= reserved_size);
     auto buf = scoped_device_block_aligned_ptr_t<ser_buffer_t>(reserved_size);
     memcpy(buf.get(), copyee, amount_to_copy);
@@ -44,8 +44,8 @@ void buf_ptr_t::resize_fill_zero(block_size_t new_size) {
     guarantee(new_size.ser_value() != 0);
     guarantee(ser_buffer_.has());
 
-    uint32_t old_reserved = compute_aligned_block_size(block_size_);
-    uint32_t new_reserved = compute_aligned_block_size(new_size);
+    uint16_t old_reserved = compute_aligned_block_size(block_size_);
+    uint16_t new_reserved = compute_aligned_block_size(new_size);
 
     if (old_reserved == new_reserved) {
         if (new_size.ser_value() < block_size_.ser_value()) {
@@ -70,8 +70,8 @@ void buf_ptr_t::resize_fill_zero(block_size_t new_size) {
 void buf_ptr_t::fill_padding_zero() const {
     guarantee(has());
     char *p = reinterpret_cast<char *>(ser_buffer());
-    uint32_t ser_block_size = block_size().ser_value();
-    uint32_t aligned = aligned_block_size();
+    uint16_t ser_block_size = block_size().ser_value();
+    uint16_t aligned = aligned_block_size();
     memset(p + ser_block_size, 0, aligned - ser_block_size);
 }
 
@@ -79,9 +79,9 @@ void buf_ptr_t::fill_padding_zero() const {
 void buf_ptr_t::assert_padding_zero() const {
     guarantee(has());
     char *p = reinterpret_cast<char *>(ser_buffer());
-    uint32_t ser_block_size = block_size().ser_value();
-    uint32_t aligned = aligned_block_size();
-    for (uint32_t i = ser_block_size; i < aligned; ++i) {
+    uint16_t ser_block_size = block_size().ser_value();
+    uint16_t aligned = aligned_block_size();
+    for (uint16_t i = ser_block_size; i < aligned; ++i) {
         rassert(p[i] == 0);
     }
 }
